+++ tsdoc/src/nodes/index.ts
@@ -1,7 +1,7 @@
 
 export * from './DocBlock';
 export * from './DocBlockTag';
-export * from './DocCodeFence';
+export * from './DocFencedCode';
 export * from './DocCodeSpan';
 export * from './DocComment';
 export * from './DocDeclarationReference';